Title: 쪼개어 배낭 채우기 구현, Time: 57ms, Memory: 2MB, Status: Wrong Answer - CodeTree

+++ 240901/쪼개어 배낭 채우기 구현/implement-fractional-knapsack.cpp
@@ -6,7 +6,7 @@
 using namespace std;
 
 int n;
-float m;
+double m;
 vector<vector<float>> v;
 
 int main() {
